+++ cccm-frontend/src/components/crna-cmp/crnaCmpFormStaticContent.vue
