+++ cccm-frontend/src/components/crna-cmp/sampleStatics.json
@@ -1,11 +1,5 @@
 {
   "display": "form",
-  "settings": {
-      "pdf": {
-          "id": "1ec0f8ee-6685-5d98-a847-26f67b67d6f0",
-          "src": "https://files.form.io/pdf/5692b91fd1028f01000407e3/file/1ec0f8ee-6685-5d98-a847-26f67b67d6f0"
-      }
-  },
   "components": [
       {
           "title": "Client Details",
