+++ cccm-frontend/src/views/CrnaCmpView.vue
@@ -3,15 +3,6 @@
 </template>
 
 <script setup lang="ts">
-    import CrnaCmpForm from '@/components/crnaCmpForm.vue'
+    import CrnaCmpForm from '@/components/crna-cmp/crnaCmpForm.vue'
 </script>
 
-<style>
-@media screen and (min-width: 768px) {
-  .about {
-    min-height: 100vh;
-    display: flex;
-    align-items: center;
-  }
-}
-</style>
